+++ Fabric/src/main/java/fuzs/puzzleslib/client/core/FabricClientFactories.java
@@ -1,8 +1,5 @@
 package fuzs.puzzleslib.client.core;
 
-import fuzs.puzzleslib.client.gui.screens.FabricScreens;
-import fuzs.puzzleslib.client.gui.screens.Screens;
-
 import java.util.function.Consumer;
 
 /**
@@ -13,10 +10,5 @@ public class FabricClientFactories implements ClientFactories {
     @Override
     public Consumer<ClientModConstructor> clientModConstructor(String modId) {
         return constructor -> FabricClientModConstructor.construct(modId, constructor);
-    }
-
-    @Override
-    public Screens screens() {
-        return FabricScreens.INSTANCE;
     }
 }