+++ Forge/src/main/java/fuzs/puzzleslib/config/ForgeConfigHolderImplV2.java
@@ -35,7 +35,7 @@ public class ForgeConfigHolderImplV2 implements ConfigHolderV2.Builder {
     }
 
     @Override
-    public <T extends AbstractConfig> Builder client(Class<T> clazz, Supplier<T> clientConfig) {
+    public <T extends AbstractConfig> Builder clientConfig(Class<T> clazz, Supplier<T> clientConfig) {
         // this is necessary to allow safely using client-only classes in the client configs (e.g. certain enums for vanilla game options)
         Supplier<T> config = () -> DistTypeExecutor.getWhenOn(DistType.CLIENT, () -> clientConfig);
         if (this.configsByClass.put(clazz, new ForgeConfigDataHolderImplV2<T>(ModConfig.Type.CLIENT, config)) != null) {
@@ -45,7 +45,7 @@ public class ForgeConfigHolderImplV2 implements ConfigHolderV2.Builder {
     }
 
     @Override
-    public <T extends AbstractConfig> Builder common(Class<T> clazz, Supplier<T> commonConfig) {
+    public <T extends AbstractConfig> Builder commonConfig(Class<T> clazz, Supplier<T> commonConfig) {
         if (this.configsByClass.put(clazz, new ForgeConfigDataHolderImplV2<T>(ModConfig.Type.COMMON, commonConfig)) != null) {
             throw new IllegalStateException(String.format("Duplicate registration for common config of type %s", clazz));
         }
@@ -53,7 +53,7 @@ public class ForgeConfigHolderImplV2 implements ConfigHolderV2.Builder {
     }
 
     @Override
-    public <T extends AbstractConfig> Builder server(Class<T> clazz, Supplier<T> serverConfig) {
+    public <T extends AbstractConfig> Builder serverConfig(Class<T> clazz, Supplier<T> serverConfig) {
         if (this.configsByClass.put(clazz, new ForgeConfigDataHolderImplV2<>(ModConfig.Type.SERVER, serverConfig)) != null) {
             throw new IllegalStateException(String.format("Duplicate registration for server config of type %s", clazz));
         }